backend/requirements.txt
# Kartavya SIEM Assistant - Backend Dependencies

# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# SIEM Connectors
elasticsearch==8.11.0
aiohttp==3.9.1

# NLP & ML
spacy==3.7.2
dateparser==1.2.0
pandas==2.1.4
numpy==1.26.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2

# Utils
redis==5.0.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
